pyproject.toml
[project]
name = "portfolio-optimizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.136.1",
    "fastmcp>=3.2.4",
    "pydantic>=2.13.3",
    "uvicorn>=0.46.0",
    "azure-storage-blob",
    "azure-identity",
    "openai",
    "numpy>=2.4.4",
    "scipy>=1.17.1",
    "python-dotenv>=1.2.2",
]
